credit_risk_project/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.132.0
matplotlib==3.10.8
numpy==2.3.4
pandas==2.3.3
pydantic==2.12.5
pytest==9.0.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
statsmodels==0.14.6
xgboost==3.2.0
httpx==0.28.1
uvicorn==0.41.0
kaggle==2.0.0